fix 調整
